add achievements

+++ src/services/ownRecipes.js
@@ -1,19 +1,26 @@
 const cloudinary = require("cloudinary");
-const Recipe = require("../models/recipe");
 
-const createRecipe = async (data) => {
+const { Recipe } = require("../models");
+const { NotFoundError } = require("../helpers/errors");
+
+const create = async (data) => {
   const recipe = await Recipe.create(data);
   return recipe;
 };
 
-const deleteRecipe = async (id, owner) => {
+const deleteById = async (id, owner) => {
   const recipe = await Recipe.findById(id);
-  await cloudinary.v2.uploader.destroy(recipe.cloudinaryImageName, "image");
+  if (!recipe) {
+    throw new NotFoundError();
+  }
+  if (recipe.cloudinaryImageName) {
+    await cloudinary.v2.uploader.destroy(recipe.cloudinaryImageName, "image");
+  }
   const result = await Recipe.findOneAndRemove({ _id: id, owner });
   return result;
 };
 
-const getRecipes = async (owner, page, limit) => {
+const get = async (owner, page, limit) => {
   const pipeline = [
     {
       $match: { owner },
@@ -33,13 +40,12 @@ const getRecipes = async (owner, page, limit) => {
       },
     },
   ];
-
   const results = await Recipe.aggregate(pipeline);
   return results[0];
 };
 
 module.exports = {
-  createRecipe,
-  deleteRecipe,
-  getRecipes,
+  create,
+  deleteById,
+  get,
 };
